Gacha Egg Mechanic
Initial implementation of the "Gacha Egg" mechanic.

This is a modified version of the Odd Egg event which allows the player to "buy" more eggs with Egg Tickets obtained from the Game Corner.

These eggs have the other Johto starters as well as a handful of other Mons. All mons will either have an exclusive move that can only be obtained via this method, or will be shiny.